backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.29
psycopg[binary]==3.1.19
python-multipart==0.0.9
pydantic-settings==2.2.1
celery==5.3.6
redis==5.0.4
boto3==1.34.116
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
PyYAML==6.0.1
Pillow==10.3.0
requests==2.32.3
playwright==1.44.0
transformers==4.46.0
accelerate==0.33.0
torch==2.3.1
torchvision==0.18.1
# Optional: for local face transfer fallback
insightface==0.7.3
opencv-python==4.10.0.84
onnxruntime-gpu==1.16.3
# For better model handling
qwen-vl-utils==0.0.2
# Logging
python-json-logger==2.0.7
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
bitsandbytes==0.42.0
